refactor FinalInput into HOC and implemented (bing change)
