+++ src/componentAtoms/FinalForm/index.js
@@ -1,5 +1,4 @@
-import FinalInput from 'componentAtoms/FinalForm/FinalInput'
-import FinalForm from 'componentAtoms/FinalForm/FinalForm'
+import { FinalForm } from 'componentAtoms/FinalForm/FinalForm'
 import { FORM_ERROR } from 'final-form'
 
-export { FinalInput, FinalForm, FORM_ERROR }
+export { FinalForm, FORM_ERROR }
