+++ src/componentAtoms/FinalInput/index.js
@@ -1,0 +1,3 @@
+import { FinalInput } from 'componentAtoms/FinalInput/FinalInput'
+
+export { FinalInput }
