+++ src/componentAtoms/InputSelect/index.js
@@ -1,0 +1,6 @@
+import { InputSelect } from 'componentAtoms/InputSelect/InputSelect'
+import { FinalInput } from 'componentAtoms/FinalInput'
+
+const InputSelectFinal = FinalInput(InputSelect)
+
+export { InputSelect, InputSelectFinal }
