+++ src/componentAtoms/InputText/index.js
@@ -1,0 +1,6 @@
+import { InputText } from 'componentAtoms/InputText/InputText'
+import { FinalInput } from 'componentAtoms/FinalInput'
+
+const InputTextFinal = FinalInput(InputText)
+
+export { InputText, InputTextFinal }
