+++ src/componentAtoms/TabPane/AccountSettingsTabPane.jsx
@@ -4,7 +4,8 @@ import { Button, FormGroup, Row, Col, Form } from 'reactstrap'
 // loader
 import Loader from 'react-loader-spinner'
 // form validation
-import { FinalInput, FinalForm, FORM_ERROR } from 'componentAtoms/FinalForm'
+import { FinalForm, FORM_ERROR } from 'componentAtoms/FinalForm'
+import { InputTextFinal } from 'componentAtoms/InputText'
 // validation
 import {
 	signInEmailValidation,
@@ -44,7 +45,7 @@ const AccountSettingsTabPane = props => {
 									</Col>
 									<Col className='align-self-center' md='9'>
 										<FormGroup>
-											<FinalInput
+											<InputTextFinal
 												type={EMAIL}
 												name={EMAIL}
 												hideSuccess
@@ -117,7 +118,7 @@ const AccountSettingsTabPane = props => {
 									</Col>
 									<Col className='align-self-center' md='9'>
 										<FormGroup>
-											<FinalInput
+											<InputTextFinal
 												type={PASSWORD}
 												name={PASSWORD}
 												onlyShowErrorOnSubmit
